:seedling: Clean up main and hvutils
Removing test functions from main.go and removing hvutils package as it
is not needed any more. The functions that were still in used are now in
the relevant packages:seedling: Clean up main and hvutils

Removing test functions from main.go and removing hvutils package as it
is not needed any more. The functions that were still in used are now in
the relevant packages. The test functions are not found in /test/hvapi.
