+++ main.go
@@ -18,16 +18,12 @@ limitations under the License.
 package main
 
 import (
-	"context"
 	"flag"
-	"fmt"
 	"os"
 
 	infrastructurev1alpha1 "github.com/hivelocity/cluster-api-provider-hivelocity/api/v1alpha1"
 	"github.com/hivelocity/cluster-api-provider-hivelocity/controllers"
-	"github.com/hivelocity/cluster-api-provider-hivelocity/pkg/hvutils"
 	hvclient "github.com/hivelocity/cluster-api-provider-hivelocity/pkg/services/hivelocity/client"
-	hv "github.com/hivelocity/hivelocity-client-go/client"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
@@ -45,8 +41,6 @@ var (
 	setupLog = ctrl.Log.WithName("setup")
 )
 
-const manualDeviceID = 14730
-
 func init() {
 	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
 	utilruntime.Must(clusterv1.AddToScheme(scheme))
@@ -55,95 +49,7 @@ func init() {
 	//+kubebuilder:scaffold:scheme
 }
 
-func cliTestListDevices(ctx context.Context, client hvclient.Client) {
-	allDevices, err := client.ListDevices(ctx)
-	if err != nil {
-		panic(err.Error())
-	}
-	device, err := hvutils.FindDeviceByTags("cn=foo", "mn=bar", allDevices)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("device: %+v\n", device)
-}
-
-func cliTestSetTags(ctx context.Context, client hvclient.Client) {
-	err := client.SetTags(ctx, manualDeviceID, []string{"foo"})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func cliTestGetDevice(ctx context.Context, client hvclient.Client) {
-	device, err := client.GetDevice(ctx, manualDeviceID)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("device: %+v\n", device)
-}
-
-func cliTestListImages(ctx context.Context, client hvclient.Client) {
-	images, err := client.ListImages(ctx, 504)
-	if err != nil {
-		panic(err)
-	}
-	for _, image := range images {
-		fmt.Println(image)
-	}
-}
-
-func cliTestListSSHKeys(ctx context.Context, client hvclient.Client) {
-	keys, err := client.ListSSHKeys(ctx)
-	if err != nil {
-		panic(err)
-	}
-	for _, sshKey := range keys {
-		fmt.Println(sshKey)
-	}
-}
-
-func cliTestProvisionDevice(ctx context.Context, client hvclient.Client) {
-	opts := hv.BareMetalDeviceUpdate{
-		Hostname: "my-host-name.example.com",
-		// Tags:     createTags("my-cluster-name", "my-host-name", false),
-		OsName:         "Ubuntu 20.x",
-		PublicSshKeyId: 861,
-	}
-	device, err := client.ProvisionDevice(ctx, manualDeviceID, opts)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("device: %+v\n", device)
-}
-
-func manualTests() {
-	factory := hvclient.HivelocityFactory{}
-	client := factory.NewClient(os.Getenv("HIVELOCITY_API_KEY"))
-	ctx := context.Background()
-	switch arg := os.Args[1]; arg {
-	case "ListDevices":
-		cliTestListDevices(ctx, client)
-		os.Exit(0)
-	case "ListSSHKeys":
-		cliTestListSSHKeys(ctx, client)
-		os.Exit(0)
-	case "ListImages":
-		cliTestListImages(ctx, client)
-		os.Exit(0)
-	case "SetTags":
-		cliTestSetTags(ctx, client)
-		os.Exit(0)
-	case "GetDevice":
-		cliTestGetDevice(ctx, client)
-		os.Exit(0)
-	case "ProvisionDevice":
-		cliTestProvisionDevice(ctx, client)
-		os.Exit(0)
-	}
-}
-
 func main() {
-	manualTests()
 	var metricsAddr string
 	var enableLeaderElection bool
 	var probeAddr string
@@ -163,24 +69,14 @@ func main() {
 	ctx := ctrl.SetupSignalHandler()
 
 	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme:                     scheme,
-		MetricsBindAddress:         metricsAddr,
-		Port:                       9443,
-		HealthProbeBindAddress:     probeAddr,
-		LeaderElection:             enableLeaderElection,
-		LeaderElectionID:           "hivelocity.cluster.x-k8s.io",
-		LeaderElectionResourceLock: "leases",
-		// LeaderElectionReleaseOnCancel defines if the leader should step down voluntarily
-		// when the Manager ends. This requires the binary to immediately end when the
-		// Manager is stopped, otherwise, this setting is unsafe. Setting this significantly
-		// speeds up voluntary leader transitions as the new leader don't have to wait
-		// LeaseDuration time first.
-		//
-		// In the default scaffold provided, the program ends immediately after
-		// the manager stops, so would be fine to enable this option. However,
-		// if you are doing or is intended to do any operation such as perform cleanups
-		// after the manager stops then its usage might be unsafe.
-		// LeaderElectionReleaseOnCancel: true,
+		Scheme:                        scheme,
+		MetricsBindAddress:            metricsAddr,
+		Port:                          9443,
+		HealthProbeBindAddress:        probeAddr,
+		LeaderElection:                enableLeaderElection,
+		LeaderElectionID:              "hivelocity.cluster.x-k8s.io",
+		LeaderElectionResourceLock:    "leases",
+		LeaderElectionReleaseOnCancel: true,
 	})
 	if err != nil {
 		setupLog.Error(err, "unable to start manager")
